Don't destroy match dialog before reading contents
This was never correct, but it worked in previous versions of gtk+ because a
reference to the Entry was held. Gtk+-2.20 stores the text in an EntryBuffer
object, which was getting unreffed and collected before ccsm could read the
text.

This fixes a bug where there is a blank match string after using the '+'
button to add a match by clicking on a window.