X11 Screen related fixes
Avoid deprecated GdkDisplay::get_number() and find screens on Gtk3 too.
Note that in order for screens to be found on Gtk3 one has to have python-Xlib.
There's no warning either way because X11 screens is an extremely marginal
feature these days.

Additionally move CurrentScreenNum from Constants to Utils: it's not that
constant.